pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Maicon de Menezes <maicondmenezes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
django-rest-swagger = "^2.2.0"
pyyaml = "^6.0.1"
markdown = "^3.6"
pygments = "^2.17.2"
uritemplate = "^4.1.1"
inflection = "^0.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.12.2"
ipykernel = "^6.29.4"
ipython = "^8.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint-review = "blue --check --diff . && isort --check --diff ."
lint = "blue . && isort ."
pre_test = "task lint"
test = "cd core && coverage run --source=. manage.py test"
post_test = "coverage html"
runserver = "python core/manage.py runserver"
makemigrations = "python core/manage.py makemigrations"
migrate = "python core/manage.py migrate"
createsuperuser = "python manage.py createsuperuser"

